+++ Tests/ChessKitTests/GameTests.swift
@@ -8,22 +8,61 @@ import XCTest
 
 class GameTests: XCTestCase {
     
-    func testAnnotation() {
-        let game = Game(
-            pgn: """
-                1. e4 e5 2. Nf3 $2 Nc6 3. Bb5 a6
-                4. Ba4 Nf6 5. O-O Be7 6. Re1 b5 7. Bb3 d6 8. c3 O-O 9. h3 Nb8 10. d4 Nbd7 $4
-                11. c4 c6 12. cxb5 axb5 13. Nc3 Bb7 14. Bg5 b4 15. Nb1 h6 16. Bh4 c5 17. dxe5 Nxe4
-                """
-        )
+    func testMakeMoves() {
+        let game = Game()
+        game.make(moves: ["e4", "e5", "Nf3", "Nc6", "Bc4"])
         
-        let assessment = Move.Assessment.good
-        let comment = "This opening is called the Ruy Lopez."
-        game?.annotate(moveAt: 3, color: .black, assessment: assessment, comment: comment)
-        
-        XCTAssertEqual(game?.moves[3]?.black?.assessment, assessment)
-        XCTAssertEqual(game?.moves[3]?.black?.comment, comment)
+        XCTAssertEqual(game.moves[.init(number: 1, color: .white, variation: 0)]?.san, "e4")
+        XCTAssertEqual(game.moves[.init(number: 1, color: .black, variation: 0)]?.san, "e5")
+        XCTAssertEqual(game.moves[.init(number: 2, color: .white, variation: 0)]?.san, "Nf3")
+        XCTAssertEqual(game.moves[.init(number: 2, color: .black, variation: 0)]?.san, "Nc6")
+        XCTAssertEqual(game.moves[.init(number: 3, color: .white, variation: 0)]?.san, "Bc4")
     }
     
+    func testMoveVariations() {
+        let game = Game()
+        game.make(moves: ["e4", "e5", "Nf3", "Nc6", "Bc4"])
+        
+        // add 2. Nc3 ... variation to 2. Nf3
+        let nf3Index = MoveIndex(number: 2, color: .white, variation: 0)
+        game.make(moves: ["Nc3", "Nf6", "Bc4"], from: nf3Index.previous)
+        
+        let nc3Index = MoveIndex(number: 2, color: .white, variation: 1)
+        
+        XCTAssertEqual(game.moves[nf3Index]?.san, "Nf3")
+        XCTAssertEqual(game.moves[nc3Index]?.san, "Nc3")
+        
+        // add 2... Nc6 ... variation to 2... Nf6
+        let nf6Index = MoveIndex(number: 2, color: .black, variation: 1)
+        game.make(moves: ["Nc6", "f4"], from: nf6Index.previous)
+        
+        let nc6Index = MoveIndex(number: 2, color: .black, variation: 2)
+        
+        XCTAssertEqual(game.moves[nf6Index]?.san, "Nf6")
+        XCTAssertEqual(game.moves[nc6Index]?.san, "Nc6")
+
+        XCTAssertEqual(
+            game.moves.previousIndex(
+                for: nc6Index
+            ),
+            nc3Index
+        )
+
+        // add another variation to 2... Nf6
+        let nc6Index2 = MoveIndex(number: 2, color: .black, variation: 0)
+        game.make(moves: ["f5", "exf5"], from: nc6Index2.previous)
+        
+        let f5Index = MoveIndex(number: 2, color: .black, variation: 3)
+        
+        XCTAssertEqual(game.moves[nc6Index2]?.san, "Nc6")
+        XCTAssertEqual(game.moves[f5Index]?.san, "f5")
+
+        XCTAssertEqual(
+            game.moves.previousIndex(
+                for: f5Index
+            ),
+            nf3Index
+        )
+    }
     
 }