+++ Tests/ChessKitTests/Parsers/PGNParserTests.swift
@@ -23,13 +23,39 @@ class PGNParserTests: XCTestCase {
         
         let game = PGNParser.parse(game: pgn)
         
-        XCTAssertEqual(game?.moves.count, 43)
-        XCTAssertEqual(game?.moves[1]?.white.assessment, .blunder)
-        XCTAssertEqual(game?.moves[3]?.black?.comment, "This opening is called the Ruy Lopez.")
-        XCTAssertEqual(game?.moves[10]?.white.end, .d4)
-        XCTAssertEqual(game?.moves[18]?.black?.piece.kind, .queen)
-        XCTAssertEqual(game?.moves[18]?.black?.end, .e7)
-        XCTAssertEqual(game?.moves[36]?.white.checkState, .check)
+        // starting position + 85 ply
+        XCTAssertEqual(game?.positions.keys.count, 86)
+        
+        XCTAssertEqual(game?.moves[.init(
+            number: 1,
+            color: .white,
+            variation: 0
+        )]?.assessment, .blunder)
+        XCTAssertEqual(game?.moves[.init(
+            number: 3,
+            color: .black,
+            variation: 0
+        )]?.comment, "This opening is called the Ruy Lopez.")
+        XCTAssertEqual(game?.moves[.init(
+            number: 10,
+            color: .white,
+            variation: 0
+        )]?.end, .d4)
+        XCTAssertEqual(game?.moves[.init(
+            number: 18,
+            color: .black,
+            variation: 0
+        )]?.piece.kind, .queen)
+        XCTAssertEqual(game?.moves[.init(
+            number: 18,
+            color: .black,
+            variation: 0
+        )]?.end, .e7)
+        XCTAssertEqual(game?.moves[.init(
+            number: 36,
+            color: .white,
+            variation: 0
+        )]?.checkState, .check)
     }
     
 }